File upload & download (upload images and generate dynamic invoices))

+++ ecommerce (Course Project)/controllers/admin.js
@@ -1,26 +1,42 @@
 const Product = require('../models/product');
 const mongodb = require('mongodb');
 const { validationResult } = require('express-validator')
+fileHelper = require('../util/file');
 
 exports.getAddProducts = (req, res, next) => {
     if (!req.session.isLoggedIn)
         return res.redirect('/login')
 
-    res.render('admin/edit-product', {
-    pageTitle: 'Add Product', 
-    path: '/admin/add-product',
-    hasError: false,
-    editing: false,
-    errorMessage: [],
-    validationErrors: []
-})
+        res.render('admin/edit-product', {
+        pageTitle: 'Add Product', 
+        path: '/admin/add-product',
+        hasError: false,
+        editing: false,
+        errorMessage: [],
+        validationErrors: []
+    })
 }
 
 exports.postAddProduct = (req, res, next) => { 
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+             pageTitle: 'Add Product',
+             path: '/admin/add-product',
+             editing: false,
+             hasError: true,
+             product: {
+                 title,
+                 price, 
+                 description
+             },
+             errorMessage: "Please attach an image.",
+             validationErrors:  []
+         });
+    }
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
        return  res.status(422).render('admin/edit-product', {
@@ -30,7 +46,6 @@ exports.postAddProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title,
-                imageUrl,
                 price, 
                 description
             },
@@ -38,10 +53,11 @@ exports.postAddProduct = (req, res, next) => {
             validationErrors:  errors.array()
         });
     }
+    const imageUrl = image.path;
    const product = new Product({title: title, 
     price: price, 
+    imageUrl,
     description: description, 
-    imageUrl: imageUrl,
     userId: req.session.user
 
 });
@@ -99,7 +115,7 @@ exports.getEditProduct = (req, res, next) => {
         res.render('admin/edit-product', {
             pageTitle: 'Edit Product',
             path: '/admin/edit-product',
-            editing:editMode,
+            editing: editMode,
             product: product,
             hasError: false, 
             errorMessage: [],
@@ -115,52 +131,55 @@ exports.getEditProduct = (req, res, next) => {
 }
 
 exports.postEditProduct = (req, res, next) => {
-    
-    // construct a new product and replace th eexisting one
     const prodId = req.body.productId;
-    // this is the edit route
-    // so i can get the new values i want to store as a part of the post request
-    // because the user enters them in the form (post req body is sent to me)
-    const { title, imageUrl, price, description } = req.body;
+    const updatedTitle = req.body.title;
+    const updatedPrice = req.body.price;
+    const image = req.file;
+    const updatedDesc = req.body.description;
+  
     const errors = validationResult(req);
+  
     if (!errors.isEmpty()) {
-        console.log(errors)
-       return  res.status(422).render('admin/edit-product', {
-            pageTitle: 'Edit Product',
-            path: '/admin/edit-product',
-            editing: true,
-            hasError: true,
-            product: {
-                title,
-                imageUrl,
-                price, 
-                description,
-                _id: prodId
-            },
-            errorMessage: errors.array()[0].msg,
-            validationErrors:  errors.array()
+      return res.status(422).render('admin/edit-product', {
+        pageTitle: 'Edit Product',
+        path: '/admin/edit-product',
+        editing: true,
+        hasError: true,
+        product: {
+          title: updatedTitle,
+          price: updatedPrice,
+          description: updatedDesc,
+          _id: prodId
+        },
+        errorMessage: errors.array()[0].msg,
+        validationErrors: errors.array()
+      });
+    }
+  
+    Product.findById(prodId)
+      .then(product => {
+        if (product.userId.toString() !== req.user._id.toString()) {
+          return res.redirect('/');
+        }
+        product.title = updatedTitle;
+        product.price = updatedPrice;
+        product.description = updatedDesc;
+        // if the user uploaded a new image
+        if (image) {
+          fileHelper.deleteFile(product.imageUrl);
+          product.imageUrl = image.path;
+        }
+        return product.save().then(result => {
+          console.log('UPDATED PRODUCT!');
+          res.redirect('/admin/products');
         });
-    }
-    Product.findByIdAndUpdate(prodId, {
-        title,
-        imageUrl,
-        price,
-        description,
-    })
-    // handle success responses from .save() promise
-    .then(result => {
-        res.redirect('/admin/products')
-    })
-    // this catch block would catch errors from findByPk() promise and the .save() promise
-   .catch(err => {
-    const error = new Error(err);
-        error.httpsStatuCode = 500;
-        // passed to a special 4 arg middleware that handles errors
+      })
+      .catch(err => {
+        const error = new Error(err);
+        error.httpStatusCode = 500;
         return next(error);
-    }
-    )
-};
-
+      });
+  };
  exports.getProducts = (req, res, next) => {
     Product.find({userId: req.user._id})
     .then(products => {
@@ -176,7 +195,14 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId, userId: req.user._id})
+    Product.findById(prodId).then( product => {
+        if (!product) {
+            return next(new Error('Product not found'))
+        }
+        fileHelper.deleteFile(product.imageUrl);
+        return Product.deleteOne({_id: prodId, userId: req.user._id})
+    }
+    )
     .then( () => {
         console.log('DELETED PRODUCT!')
         res.redirect('/admin/products')
